initial attempt; testing suite and embedding demo notebook still broken
